requirements.txt
pandas==2.2.2
praw==7.7.1
python-dotenv==1.0.1
